+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # DocuSign Click Java Client Changelog
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
+## [v1.0.0] - DocuSign Click API v1-20.4.02 - 2021-02-11
+### Changed
+- Added support for version v1-20.4.02 of the DocuSign Click API.
+- Updated the SDK release version.
+
 ## [v1.0.0-BETA] - DocuSign Click API v1.0.0 - 2020-12-28
 ### Changed
 - First Beta version of DocuSign Click API.